refactor: id types from int64 to int across the project

+++ internal/repository/class_repo.go
@@ -10,7 +10,7 @@ func NewClassRepository(db *sqlx.DB) *ClassRepository {
 	return &ClassRepository{db: db}
 }
 
-func (r *ClassRepository) ClassExists(id int64) (bool, error) {
+func (r *ClassRepository) ClassExists(id int) (bool, error) {
 	var exists bool
 	err := r.db.Get(&exists, "SELECT EXISTS(SELECT 1 FROM classes WHERE id = $1)", id)
 	return exists, err
