+++ pkg/model/teacher.go
@@ -9,7 +9,7 @@ type NewTeacher struct {
 	Name          string    `json:"name" validate:"required"`
 	DateOfBirth   time.Time `json:"dateOfBirth" validate:"required"`
 	Gender        string    `json:"gender" validate:"required"`
-	SubjectID     int64     `json:"subjectId" validate:"required"`
+	SubjectID     int     `json:"subjectId" validate:"required"`
 	HireDate      time.Time `json:"hireDate" validate:"required"`
 	WorkingStatus string    `json:"workingStatus" validate:"required"`
 }
@@ -19,17 +19,17 @@ type UpdateTeacher struct {
 	Name          string    `json:"name" validate:"required"`
 	DateOfBirth   time.Time `json:"dateOfBirth" validate:"required"`
 	Gender        string    `json:"gender" validate:"required"`
-	SubjectID     int64     `json:"subjectId" validate:"required"`
+	SubjectID     int     `json:"subjectId" validate:"required"`
 	HireDate      time.Time `json:"hireDate" validate:"required"`
 	WorkingStatus string    `json:"workingStatus" validate:"required"`
 }
 
 type TeacherDetailsSubject struct {
-	ID   int64  `json:"id"`
+	ID   int  `json:"id"`
 	Name string `json:"name"`
 }
 type TeacherDetails struct {
-	ID            int64                  `json:"id" db:"id"`
+	ID            int                  `json:"id" db:"id"`
 	Username      string                 `json:"username" db:"username"`
 	Email         *string                `json:"email" db:"email"`
 	Name          string                 `json:"name" db:"name"`
