+++ pkg/model/user.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type User struct {
-	ID          int64     `json:"id" db:"id"`
+	ID          int     `json:"id" db:"id"`
 	Username    string    `json:"username" db:"username"`
 	Password    string    `json:"password" db:"password"`
 	Email       *string   `json:"email" db:"email"`
